core: select.tsx
